ECS example with Legion
Things to note:

In order for this to compile and work, you'll need to add Legion to `Cargo.toml`:

```toml
legion = { version = "0.4.0", optional = true }
```

I am new to Rust and thus not very familiar with Cargo yet, so I'm not sure how it handles dependencies for `examples`. It'd be nicer if the dependency could only be added to the scope of this example, instead of being available to anyone using the library; Even as I made it optional, it could imply a broader support of Tetra for Legion, which is obviously not the intention here.

I based myself off `bunnymark.rs`, making small changes along the way. I never intended it to be a full conversion of that example to use ECS/Legion. However, if you feel that it's confusing as it is, I can always make it match, or contribute a different ECS example altogether. I'm happy either way!